Huge refactor: move reporting functionality to its own folder
