+++ detect_secrets/core/constants.py
@@ -1,7 +1,4 @@
 from enum import Enum
-from typing import List
-from typing import NamedTuple
-from typing import TypedDict
 
 
 # We don't scan files with these extensions.
@@ -55,33 +52,4 @@ POTENTIAL_SECRET_DETECTED_NOTE = (
     ' actual secret, and remediate it.\nOnce the secret has been'
     ' removed from the file, and another scan has been run,'
     ' its entry will be removed from the baseline file.'
-)
-
-
-class ReportSecretType(Enum):
-    AUDITED_REAL = 'Audited as real'
-    UNAUDITED = 'Unaudited'
-    LIVE = 'Live'
-
-
-class ReportExitCode(Enum):
-    PASS = 0
-    FAIL = 1
-
-
-ReportStats = TypedDict(
-    'ReportStats',
-    {'reviewed': int, 'live': int, 'unaudited': int, 'audited_real': int},
-)
-
-ReportedSecret = TypedDict(
-    'ReportedSecret',
-    {'failed_condition': ReportSecretType, 'filename': str, 'line': int, 'type': str},
-)
-
-ReportJson = TypedDict('ReportJson', {'stats': ReportStats, 'secrets': List[ReportedSecret]})
-
-ReportCheckResult = NamedTuple(
-    'ReportCheckResult',
-    [('report_exit_code', ReportExitCode), ('secrets_failing_condition', List[ReportedSecret])],
 )
