--- detect_secrets/core/reporting.py
@@ -1,1 +1,0 @@
-# TODO: move relevant functions out of audit.py to here if possible
