+++ tests/core/report/output_test.py
@@ -1,0 +1,6 @@
+# TODO: implement
+# def test_print_stats():
+#     return 0
+# TODO: implement
+# def test_report_table():
+#     return 0
